Move config to parameters as named vector
Previously, --config parameters were parsed as 'normal' doubledashed
parameters in gdal_cmd_builder. However, this --config options must be
provided as key value pairs like `gdalinfo /some/image.tif --config
GDAL_CACHEMAX "30%" for setting the maximum cache to 30 percent of the
system memory.

This commit introduces changes to the creation of the gdal commands in
gdal_cmd_builder.R to introduce the general functionality. Further, all
functions which already incorporate a config parameter were adjusted to
make use of the new functionality.

In the case of gdalinfo the "config" parameter was moved from character
parameters to repeatable parameters, which is in line with the handeling
in all other functions.